Extract lighting presets to constants and update ColorPicker and LightingSelector components
Moved lighting presets to src/constants/lighting-presets.ts and updated ColorPicker and LightingSelector to use these constants.
