+++ src/constants/lighting-presets.ts
@@ -1,0 +1,8 @@
+export const LIGHTING_PRESETS = {
+  normal: 'Normal',
+  morning: 'Morning Sun',
+  afternoon: 'Afternoon Sun',
+  evening: 'Evening',
+  night: 'Night Lights (LED)',
+  cloudy: 'Cloudy Day',
+};
